TraficViz
The first is to spot the exact site of the accident with the help of GPS and it will also determine the
seriousness of the occurred accident via accelerometer. The accident information will be sent to
the nearby police station and medical assistance station through an android mobile application with
the exact map location and severity of the accident.
The goal of this plan is to reduce traffic delays and ensure that emergency vehicles can move
freely. The idea behind this plan is to automatically green the traffic light in the ambulance's path
so that the ambulance may arrive on time and save lives. The ambulance is located by mail on the
main server's computer. Meanwhile, it regulates the traffic signal following the ambulance's
location, allowing it to reach the hospital promptly. A fully automated system locates emergency
vehicles and manages traffic signals, allowing patients to get to the hospital on time.

4.1 Accident Detection
4.1.1 Materials
• Lithium-ion rechargeable battery
• Vibration sensor
• Push-button
• LED
• Arduino
• Accelerometer
• GPS Module
• WIFI module
4.1.2 Accelerometer MPU 6050
An accelerometer sensor is used for tilt angle detection.
The "accelerometer sensor" is crucial. Fig. 5 illustrates the sensor that is used in this application.
A three-axis sensor accelerometer sensor was used that can measure acceleration (MPU 6050).
Figure out how to tilt, tap, shake, etc. the three directions of the are shown in Figure 6,7,8. tilt of
the accelerometer One of the benefits of this sensor is that it can detect the three orientations of
physical conditions
Figure 5: Accelerometer
 16 | P a g e
Figure 7. a. when vehicle position in x-axis
Figure 6: when the vehicle poistion in z-axis
4.1.3 GPS NEO 6M
Figure 8: when the vehicle position in y-axis
From “Global positioning system (GPS)”, The coded information will be transmitted through 24
satellites. Every 12 hours, these 24 satellites will circle the globe once. Information regarding
velocity, time, etc. Using GPS, we'll be able to determine how far apart two points on the globe
are, as well as the best path to take to get there.
Figure 9: GPS Module
There are three different segments in GPS they are:
a. Space segment
b. Control segment
17 | P a g e
c. User segment
For the GPS to distinguish between the different signals received by the multiple satellites, each
satellite will have a distinct code and broadcast information at different frequencies. A satellite's
distance from a GPS receiver is calculated by multiplying its journey time by its light speed. The
distance between a satellite and a GPS receiver is then calculated by multiplying its travel time by
its light speed as the satellite approaches the GPS, the control segment identifies the spacecraft
and guides it to the correct orbit and time required for the satellite to reach the GPS.
4.1.4 WIFI Module
With complete TCP/IP capabilities and an integrated MCU (Multipoint Control Unit), this device
gives the capacity to control I/O digital pins using simple pseudocode comparable to programming
language. Despite its low cost, this module has a large and expanding community.
Figure 10: WIFI Module
4.1.5 Working on the Accident detection system
To detect an accident, there are two essential modules. The first module determines if the car has
toppled over. Rather, it is affixed to the vehicle. Contains an accelerometer, Arduino
microcontroller, and WIFI module. Once a vehicle's fall is recognized, the information is sent to
the second module
Arduino microcontroller, push button, and WIFI module makes up the second module. Once the
fall has been discovered, the driver's condition will be examined, and if any abnormalities are
found, the motorist is deemed to have been in a major accident. To inform a medical center, the
Accident Detection System's buzzer must be activated. If the accident was minor the driver will
have a time of 1 minute to press the push button to stop the that is being sent to emergency numbers.
The Android application in the Smartphone will look for the nearest medical center and send a
message regarding the accident site and the accident victim's data. In addition, the information will
be sent to the emergency contact numbers via the mobile app.
As part of this project, the number of accidents will be tracked. The bike will fall to the left or right
in the event of an accident if the perpendicular axis is used as a reference. In this case, an
accelerometer is utilized to determine the vehicle's tilt. The X-axis of the accelerometer is used to
18 | P a g e
determine if the bike has fallen or not, according to the manufacturer. After a certain point, the
output voltage of an accelerometer is reduced to a low value. The microcontroller will certify that
an accident has happened when a low signal output is created and sent to it. In the car, the
accelerometer fall detecting module is installed.
4.2 Ambulance Detection
4.2.1 Materials
• Arduino nano
• RF sensors
• Traffic lights
• IR sensors
4.2.2 RF sensors (Transmitters and Receivers)
In general, wireless systems designers are constrained by two factors: they must be able to function
over a given distance and transfer a specific quantity of information within a certain data rate,
among others. Modules have a wide working voltage range of 3V to 12V and are extremely
compact.
It's 433 MHz transmitter and receiver modules that are used in the radio frequency modules. In
battery operation, the transmitter uses very little power when sending logic zero and entirely
suppressing the carrier frequency. Carriers are completely on at 4.5mA when logic one is sent with
a 3volt power source. A serial number is transferred from the transmitter to the tuned receiver,
which receives it. Two microcontrollers are used to transport data between the transmitter and
receiver.
Figure 11: RF Sensors
19 | P a g e
The performance of an RF module is influenced by several parameters compared to other radio
frequency devices. For example, by raising the transmitter's power, a wide communication range
may be achieved. To compensate for the increased electrical power consumption, battery-powered
gadgets will have a shorter operational life. The use of increased transmit power will also cause
interference with other radio frequency equipment.
4.2.3 IR Sensors
When an emergency vehicle crosses over the signal, all four IR sensors that track the signals are
activated.
When it detects an emergency vehicle crossing the IR line, the IR sensor transmits a message to
the traffic controller, causing the LED to light.
4.2.4 Working on the Ambulance Detection System
The LCD Display, IR Sensors, Signal Circuit, and GSM Modem are all connected to the Arduino
Microcontroller, which is the heart of the system.
Infrared sensors are available for Low traffic density (IR-1), Medium traffic density (IR-2), and
High-traffic density (IR-3). LCD Display shows Low traffic density and green light for 20 seconds
when the IR-1 sensor is detected. A Medium traffic density is displayed on the LCD Display when
the IR-2 sensor is detected for 40 seconds. A High Traffic Density LCD Display and Green Light
will be illuminated for 60 seconds when the IR-3 sensor is detected.
The time of red, yellow, and green traffic lights is 10sec, 1.5sec, and 10sec, respectively. In the
case of emergency when the ambulance (with IR sensor fitted) will reach the traffic signals (fitted
with IR sensors) the ambulance driver will switch on his IR sensor and the signal will be detected
by the IR sensor on the traffic light and the green light will be turned on for 3 secs. This is done
by transmitting signals from an Ambulance to a receiver located at the traffic light intersection
when it is approaching. For the Ambulance to pass through the signal junction, the green light will
be on for 60 seconds, regardless of whether there is low, medium, or high traffic density.